minor changes, update_user and create_user functions refactored, db healthcheck added to docker-compose.yml

+++ app/database/users.py
@@ -1,8 +1,9 @@
+from http import HTTPStatus
 from typing import Iterable, Type, List, Optional
 
 from fastapi import HTTPException
 
-from app.models.User import User
+from app.models.User import User, UserUpdate
 from sqlmodel import Session, select
 from app.database.engine import engine
 
@@ -12,10 +13,11 @@ def get_user(user_id: int) -> User | None:
         return session.get(User, user_id)
 
 
-def get_users() -> Iterable[User]:
+def get_users() -> List[User]:
     with Session(engine) as session:
         statement = select(User)
-        return session.exec(statement).all()
+        result = session.exec(statement).all()
+        return list(result)
 
 
 def create_user(user: User) -> User:
@@ -26,13 +28,14 @@ def create_user(user: User) -> User:
         return user
 
 
-def update_user(user_id: int, user: User) -> Type[User]: # need to fix
+def update_user(user_id: int, user: UserUpdate) -> User: # need to fix
     with Session(engine) as session:
-        db_user = session.get(User, user_id)
+        db_user: User | None = session.get(User, user_id)
         if not db_user:
-            raise  HTTPException(status_code=404, detail="User not found")
+            raise HTTPException(status_code=HTTPStatus.NOT_FOUND, detail="User not found")
         user_data = user.model_dump(exclude_unset=True)
-        db_user.sqlmodel_update(user_data)
+        for key, value in user_data.items():
+            setattr(db_user, key, value)
         session.add(db_user)
         session.commit()
         session.refresh(db_user)
@@ -42,5 +45,7 @@ def update_user(user_id: int, user: User) -> Type[User]: # need to fix
 def delete_user(user_id: int):
     with Session(engine) as session:
         user = session.get(User, user_id)
+        if not user:
+            raise HTTPException(status_code=HTTPStatus.NOT_FOUND, detail="User not found")
         session.delete(user)
         session.commit()
